move controller code to service
